requirements.txt
Django==2.2.6
djangorestframework==3.10.3
drf-writable-nested==0.5.1
odict==1.7.0
pytz==2019.3
sqlparse==0.3.0
